zapocitavani creditu podle tabulky price
git-svn-id: file:///data/svn/enum/trunk/cr/ccReg@1035 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ dbsql.h
@@ -36,7 +36,7 @@ bool DeleteFromTableMap(char *map ,int  id , int contactid );
 
 
 // vraci castku za operaci
-int GetPrice(   int action  ,  int zone , int period  );
+long GetPrice(   int action  ,  int zone , int period  );
 // zpracovani creditu
 bool UpdateCredit( int regID ,   int action  , int zone ,  int period  );
